Extract usePublishedProductsCount hook and use it in about-the-editor-menu-item
