+++ plugins/woocommerce/changelog/fix-guide-new-user
@@ -1,0 +1,5 @@
+Significance: patch
+Type: fix
+Comment: Consider if user is new or not when clicking in "About the editor"
+
+
